src/main/java/com/yura/repair/command/admin/RejectOrderCommand.java
package com.yura.repair.command.admin;

import com.yura.repair.command.Command;
import com.yura.repair.dto.OrderDto;
import com.yura.repair.service.OrderService;

import javax.servlet.http.HttpServletRequest;

public class RejectOrderCommand implements Command {

    private final OrderService orderService;

    public RejectOrderCommand(OrderService orderService) {
        this.orderService = orderService;
    }

    @Override
    public String execute(HttpServletRequest request) {
        String reason = request.getParameter("reason");
        int orderId = Integer.parseInt(request.getParameter("orderId"));
        OrderDto orderDto = orderService.findById(orderId);

        orderService.rejectOrder(orderDto, reason);

        request.getSession().setAttribute("successMessage", "reject.success");

        return "redirect:admin?command=adminOrderDetails&orderId=" + orderId;
    }
}
